// src/services/openapi/credit-checking-service/api.ts
export interface GenericBucketRequestDtoListSummaryAttachmentRequestDto {
    /**
     * 
     * @type {PageRequestDto}
     * @memberof GenericBucketRequestDtoListSummaryAttachmentRequestDto
     */
    page?: PageRequestDto;
    /**
     * 
     * @type {SortRequestDto}
     * @memberof GenericBucketRequestDtoListSummaryAttachmentRequestDto
     */
    sortList?: SortRequestDto;
    /**
     * 
     * @type {SearchDetailRequestDto}
     * @memberof GenericBucketRequestDtoListSummaryAttachmentRequestDto
     */
    searchDetail?: SearchDetailRequestDto;
    /**
     * 
     * @type {ListSummaryAttachmentRequestDto}
     * @memberof GenericBucketRequestDtoListSummaryAttachmentRequestDto
     */
    filter?: ListSummaryAttachmentRequestDto;
}
// ...
export interface ListSummaryAttachmentRequestDto {
    /**
     * 
     * @type {string}
     * @memberof ListSummaryAttachmentRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof ListSummaryAttachmentRequestDto
     */
    documentGroup?: string;
}
// ...
export interface PageRequestDto {
    /**
     * 
     * @type {number}
     * @memberof PageRequestDto
     */
    noPage?: number;
    /**
     * 
     * @type {number}
     * @memberof PageRequestDto
     */
    itemPerPage?: number;
}
// ...
export interface RequestByIdDtoLong {
    /**
     * 
     * @type {number}
     * @memberof RequestByIdDtoLong
     */
    id?: number;
}
/**
 * 
 * @export
 * @interface RequestByProcessIdDtoString
 */
export interface RequestByProcessIdDtoString {
    /**
     * 
     * @type {string}
     * @memberof RequestByProcessIdDtoString
     */
    bucketProcessId?: string;
}
// ...
export interface SearchDetailRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SearchDetailRequestDto
     */
    key?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchDetailRequestDto
     */
    value?: string;
}
// ...
export interface SortRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SortRequestDto
     */
    columnName?: string;
    /**
     * 
     * @type {string}
     * @memberof SortRequestDto
     */
    sortType?: string;
}
/**
 * 
 * @export
 * @interface SummaryAttachmentRequestDto
 */
export interface SummaryAttachmentRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SummaryAttachmentRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof SummaryAttachmentRequestDto
     */
    documentGroup?: SummaryAttachmentRequestDtoDocumentGroupEnum;
    /**
     * 
     * @type {Array<number>}
     * @memberof SummaryAttachmentRequestDto
     */
    attachments?: Array<number>;
}

/**
    * @export
    * @enum {string}
    */
export enum SummaryAttachmentRequestDtoDocumentGroupEnum {
    DIGITALMEMO = 'DIGITAL_MEMO',
    FINANCINGDOCUMENT = 'FINANCING_DOCUMENT',
    SUPPORTINGDOCUMENT = 'SUPPORTING_DOCUMENT'
}
// ...
export class SummaryControllerApi extends BaseAPI {
    /**
     * 
     * @summary API to delete summary attachment modify from database
     * @param {RequestByIdDtoLong} requestByIdDtoLong 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof SummaryControllerApi
     */
    public deleteSummaryAttachment(requestByIdDtoLong: RequestByIdDtoLong, options?: any) {
        return SummaryControllerApiFp(this.configuration).deleteSummaryAttachment(requestByIdDtoLong, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get summary by bucket process id from database
     * @param {RequestByProcessIdDtoString} requestByProcessIdDtoString 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof SummaryControllerApi
     */
    public getDetailSummary(requestByProcessIdDtoString: RequestByProcessIdDtoString, options?: any) {
        return SummaryControllerApiFp(this.configuration).getDetailSummary(requestByProcessIdDtoString, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API get list summary attachment from database
     * @param {GenericBucketRequestDtoListSummaryAttachmentRequestDto} genericBucketRequestDtoListSummaryAttachmentRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof SummaryControllerApi
     */
    public getListSummaryAttachment(genericBucketRequestDtoListSummaryAttachmentRequestDto: GenericBucketRequestDtoListSummaryAttachmentRequestDto, options?: any) {
        return SummaryControllerApiFp(this.configuration).getListSummaryAttachment(genericBucketRequestDtoListSummaryAttachmentRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to save summary by bucket process id to database
     * @param {string} [bucketProcessId] 
     * @param {any} [disclaimer] 
     * @param {any} [notes] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof SummaryControllerApi
     */
    public saveCreditCheckingSummary(bucketProcessId?: string, disclaimer?: any, notes?: any, options?: any) {
        return SummaryControllerApiFp(this.configuration).saveCreditCheckingSummary(bucketProcessId, disclaimer, notes, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to submit summary attachment modify to database
     * @param {SummaryAttachmentRequestDto} summaryAttachmentRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof SummaryControllerApi
     */
    public submitSummaryAttachmentModify(summaryAttachmentRequestDto: SummaryAttachmentRequestDto, options?: any) {
        return SummaryControllerApiFp(this.configuration).submitSummaryAttachmentModify(summaryAttachmentRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

}